home page structure - static
